Homework
1.	 ] Create a file with name “hw2.py” and download “nasdaq.txt”, “raven.txt”, “Frankenstein.txt” files into the same directory where your “hw2.py” code is located.

2.	 In “hw2.py” file, define a function make_length_wordcount() that:
a.	Takes input as the filename in the current directory (as a string)
b.	Returns a dictionary, which each key is word length and its value is the number of words with that length. 
i.	Words should be translated to lowercase.
ii.	For example, if the input file’s text is “Hello Python people Welcome to the world of Python”, then the dictionary should be: 
{2: 2, 3: 1, 5: 2, 6: 3, 7: 1} 

3.	 In “hw2.py” file, define a function make_word_count() that:
a.	Takes input as the filename in the current directory (as a string)
b.	Returns a dictionary, which each key is a word and its value is the number of occurrences of that word.
i.	Words should be translated to lowercase.
ii.	For example, if the input file’s text is “Hello P